Refactor almost the entire program
First of all, this could have been multiple commits, but I'd rather have it in a state that actually works before doing that.

This is a major refactor of the program. The rendering process is more flexible and uses G'MIC to properly palettize images. It fully supports (up to 3) remappable colors through the use of material indexes and can support masking and other things through multiple render layers.

Additionally, the addon will generate a JSON file from the properties that you specify in the addon panel, and combine the images and the JSON file into a .parkobj file.

Right now the only object type implemented is small scenery, but there is also a "custom" object mode that lets you directly specify render options.